backend/requirements.txt
google-cloud-bigquery[pandas]>=3.31.0
google-auth>=2.39.0
pandas>=2.2.0
db-dtypes>=1.4.0
pyarrow>=19.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
